blockRead/blockWrite is 32-bits wide instead of 16-bits for capacity

+++ include/Cache.h
@@ -284,14 +284,14 @@ public:
         // cache the first entry
         (void)getByte(0);
     }
-    size_t blockWrite(Address address, uint8_t *buf, size_t capacity) noexcept override {
+    uint32_t blockWrite(Address address, uint8_t *buf, uint32_t capacity) noexcept override {
         //return MemoryThing::blockWrite(address, buf, capacity);
         /// @todo see if it makes more sense to require that the user disables the cache if they need to do such a thing
 
         if (enabled_) {
             // use the cache where it makes sense
-            size_t numWritten = 0;
-            for (size_t i = 0; i < capacity; ++i, ++address, ++numWritten) {
+            uint32_t numWritten = 0;
+            for (uint32_t i = 0; i < capacity; ++i, ++address, ++numWritten) {
                 setByte(address, buf[i]);
             }
             return numWritten;
@@ -300,12 +300,12 @@ public:
             return thing_.blockWrite(address, buf, capacity);
         }
     }
-    size_t blockRead(Address address, uint8_t *buf, size_t capacity) noexcept override {
+    uint32_t blockRead(Address address, uint8_t *buf, uint32_t capacity) noexcept override {
         // we want to directly read from the underlying memory thing using the buffer so we need to do cache coherency checks as well
         /// @todo figure out if we actually need to disable the cache when performing a read from memory. I don't think we need to actually
         if (enabled_) {
-            size_t numRead = 0;
-            for (size_t i = 0; i < capacity; ++i, ++address, ++numRead) {
+            uint32_t numRead = 0;
+            for (uint32_t i = 0; i < capacity; ++i, ++address, ++numRead) {
                 buf[i] = getByte(address);
             }
             return numRead;
